requirements.txt
fastapi
uvicorn
together
pdfplumber
python-docx
sentence-transformers
faiss-cpu
pydantic
python-dotenv
jinja2
